more readme, added global .h

+++ Pod/Classes/EventCounterService.h
@@ -10,11 +10,11 @@
 
 @interface EventCounterService : NSObject
 
-+(NSUInteger)addCountForEvent:(NSString*)event count:(int)add;
++(NSUInteger)addCountForEvent:(NSString*)event count:(NSInteger)add;
 +(NSUInteger)addCountForEvent:(NSString*)event;
 +(void)resetCountForEvent:(NSString*)event;
 +(NSUInteger)checkCountForEvent:(NSString*)event;
 +(NSUInteger)checkDailyEventCount:(NSString*)dailyEvent;
 +(NSUInteger)addDailyEventCount:(NSString*)dailyEvent;
-+(NSUInteger)addCountForDailyEvent:(NSString*)dailyEvent count:(int)count;
++(NSUInteger)addCountForDailyEvent:(NSString*)dailyEvent count:(NSInteger)count;
 @end
